feat(eks): refactor by extracting the aws-operations packages
