+++ pkg/cluster-providers/eks/cluster.go
@@ -7,17 +7,15 @@ import (
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/service/ec2"
-	ec2Types "github.com/aws/aws-sdk-go-v2/service/ec2/types"
 	"github.com/aws/aws-sdk-go-v2/service/eks"
-	"github.com/aws/aws-sdk-go-v2/service/eks/types"
 	"github.com/aws/aws-sdk-go-v2/service/iam"
+	"github.com/kumahq/kuma-smoke/pkg/cluster-providers/eks/aws-operations"
 	err_pkg "github.com/pkg/errors"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"strings"
 	"sync"
-	"time"
 
 	"github.com/blang/semver/v4"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -55,7 +53,7 @@ func NewFromExisting(ctx context.Context, name string) (*Cluster, error) {
 		return nil, err_pkg.Wrap(err, "failed to load AWS SDK config")
 	}
 
-	restCfg, kubeCfg, err := clientForCluster(ctx, cfg, name)
+	restCfg, kubeCfg, err := aws_operations.ClientForCluster(ctx, cfg, name)
 	if err != nil {
 		return nil, err_pkg.Wrapf(err, "failed to get kube client for cluster %s", name)
 	}
@@ -127,28 +125,28 @@ func (c *Cluster) Cleanup(ctx context.Context) error {
 	}
 
 	vpcID := activeCluster.Cluster.ResourcesVpcConfig.VpcId
-	ngRole, launchTemplateId, err := deleteNodeGroup(ctx, eksClient, c.name)
+	ngRole, launchTemplateId, err := aws_operations.DeleteNodeGroup(ctx, eksClient, c.name)
 	if err != nil {
 		return err
 	}
 	if launchTemplateId != "" {
-		err = deleteNodeLaunchTemplate(ctx, ec2Client, launchTemplateId)
+		err = aws_operations.DeleteNodeLaunchTemplate(ctx, ec2Client, launchTemplateId)
 		if err != nil {
 			return err
 		}
 	}
 
-	err = deleteRoles(ctx, iamClient, []string{ngRole, *activeCluster.Cluster.RoleArn})
+	err = aws_operations.DeleteRoles(ctx, iamClient, []string{ngRole, *activeCluster.Cluster.RoleArn})
 	if err != nil {
 		return err
 	}
 
-	err = deleteCluster(ctx, eksClient, c.name)
+	err = aws_operations.DeleteCluster(ctx, eksClient, c.name)
 	if err != nil {
 		return err
 	}
 
-	err = deleteVPC(ctx, ec2Client, *vpcID)
+	err = aws_operations.DeleteVPC(ctx, ec2Client, *vpcID)
 	if err != nil {
 		return err
 	}
@@ -282,244 +280,4 @@ func (c *Cluster) DumpDiagnostics(ctx context.Context, meta string) (string, err
 
 func (c *Cluster) IPFamily() clusters.IPFamily {
 	return c.ipFamily
-}
-
-func deleteNodeGroup(ctx context.Context, eksClient *eks.Client, clusterName string) (string, string, error) {
-	var notFoundErr *types.ResourceNotFoundException
-	describeNGInput := &eks.DescribeNodegroupInput{
-		ClusterName:   aws.String(clusterName),
-		NodegroupName: aws.String(defaultNodeGroupName),
-	}
-	ngInfo, err := eksClient.DescribeNodegroup(ctx, describeNGInput)
-	if err != nil {
-		if errors.As(err, &notFoundErr) {
-			// the node group had already been deleted
-			return "", "", nil
-		} else {
-			return "", "", err_pkg.Wrapf(err, "failed to describe node group %s of cluster %s", defaultNodeGroupName, clusterName)
-		}
-	}
-
-	nodeGroupInput := &eks.DeleteNodegroupInput{
-		ClusterName:   aws.String(clusterName),
-		NodegroupName: aws.String(defaultNodeGroupName),
-	}
-	_, err = eksClient.DeleteNodegroup(ctx, nodeGroupInput)
-	if err != nil {
-		return "", "", err
-	}
-
-	ticker := time.NewTicker(5 * time.Second)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return "", "", ctx.Err()
-		case <-ticker.C:
-			describeInput := &eks.DescribeNodegroupInput{
-				ClusterName:   aws.String(clusterName),
-				NodegroupName: aws.String(defaultNodeGroupName),
-			}
-			_, err := eksClient.DescribeNodegroup(ctx, describeInput)
-			if err != nil {
-				if errors.As(err, &notFoundErr) {
-					// the node group has already been deleted successfully
-					return aws.ToString(ngInfo.Nodegroup.NodeRole), aws.ToString(ngInfo.Nodegroup.LaunchTemplate.Id), nil
-				} else {
-					return "", "", err_pkg.Wrap(err, fmt.Sprintf("failed to describe node group %s of cluster %s", defaultNodeGroupName, clusterName))
-				}
-			}
-		}
-	}
-}
-
-func deleteNodeLaunchTemplate(ctx context.Context, ec2Client *ec2.Client, launchTemplateId string) error {
-	deleteLaunchTmplInput := &ec2.DeleteLaunchTemplateInput{
-		LaunchTemplateId: aws.String(launchTemplateId),
-	}
-	_, err := ec2Client.DeleteLaunchTemplate(ctx, deleteLaunchTmplInput)
-	if err != nil {
-		return err_pkg.Wrapf(err, "failed to delete node launch template %s", launchTemplateId)
-	}
-	return nil
-}
-
-func deleteCluster(ctx context.Context, eksClient *eks.Client, clusterName string) error {
-	var notFoundErr *types.ResourceNotFoundException
-	clusterInput := &eks.DeleteClusterInput{
-		Name: aws.String(clusterName),
-	}
-	_, err := eksClient.DeleteCluster(ctx, clusterInput)
-	if err != nil {
-		return err
-	}
-
-	ticker := time.NewTicker(5 * time.Second)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-ticker.C:
-			describeInput := &eks.DescribeClusterInput{
-				Name: aws.String(clusterName),
-			}
-			_, err := eksClient.DescribeCluster(ctx, describeInput)
-			if err != nil {
-				if errors.As(err, &notFoundErr) {
-					// the cluster has already been deleted successfully
-					return nil
-				} else {
-					return err_pkg.Wrap(err, fmt.Sprintf("failed to describe EKS cluster %s", clusterName))
-				}
-			}
-		}
-	}
-}
-
-func deleteVPC(ctx context.Context, ec2Client *ec2.Client, vpcID string) error {
-	routeTablesOutput, err := ec2Client.DescribeRouteTables(ctx, &ec2.DescribeRouteTablesInput{
-		Filters: []ec2Types.Filter{
-			{Name: aws.String("vpc-id"), Values: []string{vpcID}},
-		},
-	})
-	if err != nil {
-		return err_pkg.Wrapf(err, "failed to list route tables in VPC %s", vpcID)
-	}
-
-	for _, rt := range routeTablesOutput.RouteTables {
-		isMain := false
-		for _, assoc := range rt.Associations {
-			if assoc.Main != nil && *assoc.Main {
-				isMain = true
-				break
-			}
-		}
-		if isMain {
-			continue
-		}
-
-		for _, assoc := range rt.Associations {
-			if assoc.RouteTableAssociationId != nil {
-				_, err := ec2Client.DisassociateRouteTable(ctx, &ec2.DisassociateRouteTableInput{
-					AssociationId: assoc.RouteTableAssociationId,
-				})
-				if err != nil {
-					return err_pkg.Wrapf(err, "failed to disassociate route table association %s for route table %s", *assoc.RouteTableAssociationId, *rt.RouteTableId)
-				}
-			}
-		}
-
-		_, err := ec2Client.DeleteRouteTable(ctx, &ec2.DeleteRouteTableInput{
-			RouteTableId: rt.RouteTableId,
-		})
-		if err != nil {
-			return err_pkg.Wrapf(err, "failed to delete route table %s", *rt.RouteTableId)
-		}
-	}
-
-	subnetsOutput, err := ec2Client.DescribeSubnets(ctx, &ec2.DescribeSubnetsInput{
-		Filters: []ec2Types.Filter{
-			{Name: aws.String("vpc-id"), Values: []string{vpcID}},
-		},
-	})
-	if err != nil {
-		return err_pkg.Wrapf(err, "failed to describe subnets in VPC %s", vpcID)
-	}
-
-	for _, subnet := range subnetsOutput.Subnets {
-		_, err := ec2Client.DeleteSubnet(ctx, &ec2.DeleteSubnetInput{
-			SubnetId: subnet.SubnetId,
-		})
-		if err != nil {
-			return err_pkg.Wrapf(err, "failed to delete subnet %s", *subnet.SubnetId)
-		}
-	}
-
-	igwsOutput, err := ec2Client.DescribeInternetGateways(ctx, &ec2.DescribeInternetGatewaysInput{
-		Filters: []ec2Types.Filter{
-			{Name: aws.String("attachment.vpc-id"), Values: []string{vpcID}},
-		},
-	})
-	if err != nil {
-		return err_pkg.Wrapf(err, "failed to describe internet gateways in VPC %s", vpcID)
-	}
-
-	for _, igw := range igwsOutput.InternetGateways {
-		_, err := ec2Client.DetachInternetGateway(ctx, &ec2.DetachInternetGatewayInput{
-			InternetGatewayId: igw.InternetGatewayId,
-			VpcId:             aws.String(vpcID),
-		})
-		if err != nil {
-			return err_pkg.Wrapf(err, "failed to detach internet gateway %s", *igw.InternetGatewayId)
-		}
-
-		_, err = ec2Client.DeleteInternetGateway(ctx, &ec2.DeleteInternetGatewayInput{
-			InternetGatewayId: igw.InternetGatewayId,
-		})
-		if err != nil {
-			return err_pkg.Wrapf(err, "failed to delete internet gateway %s", *igw.InternetGatewayId)
-		}
-	}
-
-	sgOutput, err := ec2Client.DescribeSecurityGroups(ctx, &ec2.DescribeSecurityGroupsInput{
-		Filters: []ec2Types.Filter{
-			{Name: aws.String("vpc-id"), Values: []string{vpcID}},
-		},
-	})
-	if err != nil {
-		return err_pkg.Wrapf(err, "failed to describe security groups in VPC %s", vpcID)
-	}
-
-	for _, sg := range sgOutput.SecurityGroups {
-		if sg.GroupName != nil && *sg.GroupName == "default" {
-			continue
-		}
-
-		for _, ingress := range sg.IpPermissions {
-			_, err := ec2Client.RevokeSecurityGroupIngress(ctx, &ec2.RevokeSecurityGroupIngressInput{
-				GroupId:       sg.GroupId,
-				IpPermissions: []ec2Types.IpPermission{ingress},
-			})
-			if err != nil {
-				return err_pkg.Wrapf(err, "failed to revoke a %s ingress rule on security group %s",
-					aws.ToString(ingress.IpProtocol), aws.ToString(sg.GroupId))
-			}
-		}
-
-		for _, egress := range sg.IpPermissionsEgress {
-			_, err := ec2Client.RevokeSecurityGroupEgress(ctx, &ec2.RevokeSecurityGroupEgressInput{
-				GroupId:       sg.GroupId,
-				IpPermissions: []ec2Types.IpPermission{egress},
-			})
-			if err != nil {
-				return err_pkg.Wrapf(err, "failed to revoke a %s egress rule on security group %s",
-					aws.ToString(egress.IpProtocol), aws.ToString(sg.GroupId))
-			}
-		}
-	}
-
-	for _, sg := range sgOutput.SecurityGroups {
-		if sg.GroupName != nil && *sg.GroupName == "default" {
-			continue
-		}
-
-		_, err := ec2Client.DeleteSecurityGroup(ctx, &ec2.DeleteSecurityGroupInput{
-			GroupId: sg.GroupId,
-		})
-		if err != nil {
-			return err_pkg.Wrapf(err, "failed to delete security group %s", *sg.GroupId)
-		}
-	}
-
-	_, err = ec2Client.DeleteVpc(ctx, &ec2.DeleteVpcInput{
-		VpcId: aws.String(vpcID),
-	})
-	if err != nil {
-		return err_pkg.Wrapf(err, "failed to delete VPC %s", vpcID)
-	}
-
-	return nil
 }
